ANDROID-386 Cannot cancel request to join sites
